if local directory is empty as well

+++ src/ec504project/application/FileObj.java
@@ -21,9 +21,11 @@ public class FileObj {
 		int index = 0;
 		
 		if(senderList.size() == 0){
-			for(int jj=fileList.size()-1; jj >= 0; jj--){				
-				System.out.println("Removing file: "+ fileList.get(jj).filePath);
-				fileList.get(jj).filePath.delete();
+			if(fileList.size() > 0){
+				for(int jj=fileList.size()-1; jj >= 0; jj--){				
+					System.out.println("Removing file: "+ fileList.get(jj).filePath);
+					fileList.get(jj).filePath.delete();
+				}
 			}
 			return DiffList;
 		}
